Move Circe-based tests to shared sources
Circe now supports Scala Native
